requirements_test.txt
aiohttp_cors>=0.7.0
codespell==2.3.0
coverage>=7.3.1
fnv-hash-fast>=0.5.0
mypy-dev==1.11.0a6
pre-commit==3.7.1
psutil-home-assistant>=0.0.1
pytest>=7.4.3
pytest-asyncio>=0.21.0
pytest-homeassistant-custom-component>=0.13.82
ruff==0.4.7
tomli==2.0.1
tox==4.15.0
